@@ -45,40 +58,58 @@
     }
   ],
   "dependencies": {
-    "async": "0.9.0",
-    "bindings": "1.2.1",
-    "debug": "^2.1.1",
-    "nan": "2.3.5",
-    "node-pre-gyp": "0.6.x",
-    "node-pre-gyp-github": "^1.1.0",
-    "optimist": "~0.6.1",
-    "sf": "0.1.7"
+    "bindings": "1.3.0",
+    "commander": "^2.11.0",
+    "debug": "^3.0.1",
+    "nan": "^2.6.2",
+    "prebuild-install": "^2.2.1",
+    "promirepl": "^1.0.1",
+    "safe-buffer": "^5.0.1"
   },
-  "bundledDependencies": [
-    "node-pre-gyp",
-    "node-pre-gyp-github"
-  ],
   "devDependencies": {
-    "mocha": "*",
-    "chai": "*",
-    "sinon-chai": "*",
-    "sinon": "*",
-    "grunt": "*",
-    "grunt-cli": "*",
-    "grunt-mocha-test": "*",
-    "grunt-contrib-jshint": "*",
-    "sandboxed-module": "~0.3.0"
+    "bluebird": "^3.5.0",
+    "chai": "^4.0.2",
+    "chai-subset": "^1.5.0",
+    "conventional-changelog-cli": "^1.3.2",
+    "eslint": "^4.7.2",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-node": "^5.1.0",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "istanbul": "^0.4.4",
+    "jsdoc-to-markdown": "^3.0.0",
+    "mocha": "^3.4.2",
+    "prebuild": "^6.2.1",
+    "proxyquire": "^1.7.10",
+    "sinon": "^3.3.0"
   },
   "engines": {
-    "node": ">= 0.10.0"
+    "node": ">=4.0.0"
   },
   "bin": {
-    "serialportlist": "./bin/serialportList.js",
-    "serialportterm": "./bin/serialportTerminal.js"
+    "serialport-list": "./bin/list.js",
+    "serialport-repl": "./bin/repl.js",
+    "serialport-term": "./bin/terminal.js"
   },
   "license": "MIT",
   "scripts": {
-    "install": "node-pre-gyp install --fallback-to-build",
-    "test": "grunt --verbose"
-  }
+    "arduino-test": "TEST_PORT=$(./bin/find-arduino.js) npm test",
+    "changelog": "conventional-changelog -i CHANGELOG.md -p angular -s",
+    "docs": "jsdoc2md --no-cache -t .docs/README.hbs --partial .docs/sig-name.hbs --partial .docs/sig-link.hbs --partial .docs/edit-warning.hbs -r table --separators --name-format -f lib/* lib/bindings/* lib/parsers/* > README.md",
+    "docs:diff": "jsdoc2md --no-cache -t .docs/README.hbs --partial .docs/sig-name.hbs --partial .docs/sig-link.hbs --partial .docs/edit-warning.hbs -r table --separators --name-format -f lib/* lib/bindings/* lib/parsers/* | diff README.md - || (echo 'Docs out of date, run `npm run docs` and commit the new README.md' && false)",
+    "lint": "eslint lib test bin examples",
+    "rebuild-all": "npm rebuild && node-gyp rebuild",
+    "repl": "node bin/repl.js",
+    "stress": "mocha --no-timeouts test/arduinoTest/stress.js",
+    "test": "istanbul cover ./node_modules/mocha/bin/_mocha",
+    "test:watch": "mocha -w",
+    "valgrind": "TEST_PORT=$(./bin/find-arduino.js) valgrind --leak-check=full node test/manual-testing/memory-binding.js",
+    "debugger": "TEST_PORT=$(./bin/find-arduino.js) node --inspect --debug-brk test/manual-testing/performance.js",
+    "install": "prebuild-install || node-gyp rebuild",
+    "rebuild": "prebuild --compile",
+    "prebuild": "prebuild --all --strip --verbose",
+    "prebuild-upload": "prebuild --all --strip --verbose"
+  },
+  "gypfile": true
 }
