@@ -1,6 +1,7 @@
-#ifndef WIN32
-#include "serialport.h"
-#include "serialport_poller.h"
+#include "./serialport_unix.h"
+#include "./serialport.h"
+
+#include <sys/file.h>
 #include <unistd.h>
 #include <fcntl.h>
 #include <errno.h>
@@ -9,19 +10,11 @@
 #ifdef __APPLE__
 #include <AvailabilityMacros.h>
 #include <sys/param.h>
-#include <IOKit/IOKitLib.h>
-#include <IOKit/IOCFPlugIn.h>
-#include <IOKit/usb/IOUSBLib.h>
-#include <IOKit/serial/IOSerialKeys.h>
-
-uv_mutex_t list_mutex;
-Boolean lockInitialised = FALSE;
 #endif
 
 #if defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4)
 #include <sys/ioctl.h>
 #include <IOKit/serial/ioss.h>
-#include <errno.h>
 #endif
 
 #if defined(__OpenBSD__)
@@ -33,31 +26,7 @@ Boolean lockInitialised = FALSE;
 #include <linux/serial.h>
 #endif
 
-struct UnixPlatformOptions : OpenBatonPlatformOptions {
-public:
-  uint8_t vmin;
-  uint8_t vtime;
-};
-
-OpenBatonPlatformOptions* ParsePlatformOptions(const v8::Local<v8::Object>& options) {
-  Nan::HandleScope scope;
-
-  UnixPlatformOptions* result = new UnixPlatformOptions();
-  result->vmin = Nan::Get(options, Nan::New<v8::String>("vmin").ToLocalChecked()).ToLocalChecked()->ToInt32()->Int32Value();
-  result->vtime = Nan::Get(options, Nan::New<v8::String>("vtime").ToLocalChecked()).ToLocalChecked()->ToInt32()->Int32Value();
-
-  return result;
-}
-
-int ToBaudConstant(int baudRate);
-int ToDataBitsConstant(int dataBits);
 int ToStopBitsConstant(SerialPortStopBits stopBits);
-
-void AfterOpenSuccess(int fd, Nan::Callback* dataCallback, Nan::Callback* disconnectedCallback, Nan::Callback* errorCallback) {
-  delete dataCallback;
-  delete errorCallback;
-  delete disconnectedCallback;
-}
 
 int ToBaudConstant(int baudRate) {
   switch (baudRate) {
@@ -98,23 +67,6 @@ int ToBaudConstant(int baudRate) {
   return -1;
 }
 
-#ifdef __APPLE__
-typedef struct SerialDevice {
-    char port[MAXPATHLEN];
-    char locationId[MAXPATHLEN];
-    char vendorId[MAXPATHLEN];
-    char productId[MAXPATHLEN];
-    char manufacturer[MAXPATHLEN];
-    char serialNumber[MAXPATHLEN];
-} stSerialDevice;
-
-typedef struct DeviceListItem {
-    struct SerialDevice value;
-    struct DeviceListItem *next;
-    int* length;
-} stDeviceListItem;
-#endif
-
 int ToDataBitsConstant(int dataBits) {
   switch (dataBits) {
     case 8: default: return CS8;
@@ -125,147 +77,119 @@ int ToDataBitsConstant(int dataBits) {
   return -1;
 }
 
-
-
 void EIO_Open(uv_work_t* req) {
   OpenBaton* data = static_cast<OpenBaton*>(req->data);
 
   int flags = (O_RDWR | O_NOCTTY | O_NONBLOCK | O_CLOEXEC | O_SYNC);
   int fd = open(data->path, flags);
 
-  if(-1 == setup(fd, data)){
+  if (-1 == fd) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot open %s", strerror(errno), data->path);
+    return;
+  }
+
+  if (-1 == setup(fd, data)) {
+    close(fd);
     return;
   }
 
   data->result = fd;
+}
+
+int setBaudRate(ConnectionOptionsBaton *data) {
+  // lookup the standard baudrates from the table
+  int baudRate = ToBaudConstant(data->baudRate);
+  int fd = data->fd;
+
+  // get port options
+  struct termios options;
+  if (-1 == tcgetattr(fd, &options)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s setting custom baud rate of %d", strerror(errno), data->baudRate);
+    return -1;
+  }
+
+  // If there is a custom baud rate on linux you can do the following trick with B38400
+  #if defined(__linux__) && defined(ASYNC_SPD_CUST)
+    if (baudRate == -1) {
+      struct serial_struct serinfo;
+      serinfo.reserved_char[0] = 0;
+      if (-1 != ioctl(fd, TIOCGSERIAL, &serinfo)) {
+        serinfo.flags &= ~ASYNC_SPD_MASK;
+        serinfo.flags |= ASYNC_SPD_CUST;
+        serinfo.custom_divisor = (serinfo.baud_base + (data->baudRate / 2)) / data->baudRate;
+        if (serinfo.custom_divisor < 1)
+          serinfo.custom_divisor = 1;
+
+        ioctl(fd, TIOCSSERIAL, &serinfo);
+        ioctl(fd, TIOCGSERIAL, &serinfo);
+      } else {
+        snprintf(data->errorString, sizeof(data->errorString), "Error: %s setting custom baud rate of %d", strerror(errno), data->baudRate);
+        return -1;
+      }
+
+      // Now we use "B38400" to trigger the special baud rate.
+      baudRate = B38400;
+    }
+  #endif
+
+  // On OS X, starting with Tiger, we can set a custom baud rate with ioctl
+  #if defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4)
+    if (-1 == baudRate) {
+      speed_t speed = data->baudRate;
+      if (-1 == ioctl(fd, IOSSIOSPEED, &speed)) {
+        snprintf(data->errorString, sizeof(data->errorString), "Error: %s calling ioctl(.., IOSSIOSPEED, %ld )", strerror(errno), speed );
+        return -1;
+      } else {
+        tcflush(fd, TCIOFLUSH);
+        return 1;
+      }
+    }
+  #endif
+
+  if (-1 == baudRate) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error baud rate of %d is not supported on your platform", data->baudRate);
+    return -1;
+  }
+
+  // If we have a good baud rate set it and lets go
+  cfsetospeed(&options, baudRate);
+  cfsetispeed(&options, baudRate);
+  // throw away all the buffered data
+  tcflush(fd, TCIOFLUSH);
+  // make the changes now
+  tcsetattr(fd, TCSANOW, &options);
+  return 1;
 }
 
 void EIO_Update(uv_work_t* req) {
-  OpenBaton* data = static_cast<OpenBaton*>(req->data);
-
-  int fd = data->fd;
-
-  if(-1 == setup(fd, data)){
-    return;
-  }
-
-  data->result = fd;
+  ConnectionOptionsBaton* data = static_cast<ConnectionOptionsBaton*>(req->data);
+  setBaudRate(data);
 }
 
-
 int setup(int fd, OpenBaton *data) {
-
-  UnixPlatformOptions* platformOptions = static_cast<UnixPlatformOptions*>(data->platformOptions);
-
-  int baudRate = ToBaudConstant(data->baudRate);
-
-// #if not ( defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4) )
-//   if(baudRate == -1) {
-//     snprintf(data->errorString, sizeof(data->errorString), "Invalid baud rate setting %d", data->baudRate);
-//     return;
-//   }
-// #endif
-
   int dataBits = ToDataBitsConstant(data->dataBits);
-  if(dataBits == -1) {
+  if (-1 == dataBits) {
     snprintf(data->errorString, sizeof(data->errorString), "Invalid data bits setting %d", data->dataBits);
     return -1;
   }
 
-
-  // int flags = (O_RDWR | O_NOCTTY | O_NONBLOCK | O_CLOEXEC | O_SYNC);
-  // if(data->hupcl == false) {
-  //   flags &= ~HUPCL;
-  // }
-  // int fd = open(data->path, flags);
-
-  int flags = (O_RDWR | O_NOCTTY | O_NONBLOCK | O_CLOEXEC | O_SYNC);
-  if(data->hupcl == false) {
-    flags &= ~HUPCL;
-  }
-  fd = open(data->path, flags);
-
-
-  if (fd == -1) {
-    snprintf(data->errorString, sizeof(data->errorString), "Cannot open %s", data->path);
+  // Snow Leopard doesn't have O_CLOEXEC
+  if (-1 == fcntl(fd, F_SETFD, FD_CLOEXEC)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error %s Cannot open %s", strerror(errno), data->path);
     return -1;
   }
 
-  //Snow Leopard doesn't have O_CLOEXEC
-  int cloexec = fcntl(fd, F_SETFD, FD_CLOEXEC);
-  if (cloexec == -1) {
-    snprintf(data->errorString, sizeof(data->errorString), "Cannot open %s", data->path);
-    return -1;
-  }
-
-  // struct sigaction saio;
-  // saio.sa_handler = sigio_handler;
-  // sigemptyset(&saio.sa_mask);
-  // saio.sa_flags = 0;
-  // sigaction(SIGIO, &saio, NULL);
-
-  // //all process to receive SIGIO
-  // fcntl(fd, F_SETOWN, getpid());
-  // int flflags = fcntl(fd, F_GETFL);
-  // fcntl(fd, F_SETFL, flflags | FNONBLOCK);
-
+  // Get port configuration for modification
   struct termios options;
-  // Set baud and other configuration.
   tcgetattr(fd, &options);
 
-// Removing check for valid BaudRates due to ticket: #140
-// #if not ( defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4) )
-  // Specify the baud rate
-
-
-  // On linux you can alter the meaning of B38400 to mean a custom baudrate...
-#if defined(__linux__) && defined(ASYNC_SPD_CUST)
-  if (baudRate == -1) {
-    struct serial_struct serinfo;
-    serinfo.reserved_char[0] = 0;
-    if (ioctl(fd, TIOCGSERIAL, &serinfo) != -1) {
-      serinfo.flags &= ~ASYNC_SPD_MASK;
-      serinfo.flags |= ASYNC_SPD_CUST;
-      serinfo.custom_divisor = (serinfo.baud_base + (data->baudRate / 2)) / data->baudRate;
-      if (serinfo.custom_divisor < 1)
-        serinfo.custom_divisor = 1;
-
-      ioctl(fd, TIOCSSERIAL, &serinfo);
-      ioctl(fd, TIOCGSERIAL, &serinfo);
-      // if (serinfo.custom_divisor * rate != serinfo.baud_base) {
-      //   warnx("actual baudrate is %d / %d = %f",
-      //     serinfo.baud_base, serinfo.custom_divisor,
-      //     (float)serinfo.baud_base / serinfo.custom_divisor);
-      // }
-    }
-
-    // Now we use "B38400" to trigger the special baud rate.
-    baudRate = B38400;
-  }
-#endif
-
-  if (baudRate != -1) {
-    cfsetispeed(&options, baudRate);
-    cfsetospeed(&options, baudRate);
-  }
-
-// Removing check for valid BaudRates due to ticket: #140
-// #endif
-
-  /*
-    IGNPAR  : ignore bytes with parity errors
-  */
+  // IGNPAR: ignore bytes with parity errors
   options.c_iflag = IGNPAR;
 
-  /*
-    ICRNL   : map CR to NL (otherwise a CR input on the other computer
-              will not terminate input)
-  */
-  // Pulling this for now. It should be an option, however. -Giseburt
-  //options.c_iflag = ICRNL;
-
-  //  otherwise make device raw (no other input processing)
-
+  // ICRNL: map CR to NL (otherwise a CR input on the other computer will not terminate input)
+  // Future potential option
+  // options.c_iflag = ICRNL;
+  // otherwise make device raw (no other input processing)
 
   // Specify data bits
   options.c_cflag &= ~CSIZE;
@@ -292,9 +216,7 @@ int setup(int fd, OpenBaton *data) {
     options.c_iflag |= IXANY;
   }
 
-
-  switch (data->parity)
-  {
+  switch (data->parity) {
   case SERIALPORT_PARITY_NONE:
     options.c_cflag &= ~PARENB;
     // options.c_cflag &= ~CSTOPB;
@@ -317,11 +239,10 @@ int setup(int fd, OpenBaton *data) {
     break;
   default:
     snprintf(data->errorString, sizeof(data->errorString), "Invalid parity setting %d", data->parity);
-    close(fd);
     return -1;
   }
 
-  switch(data->stopBits) {
+  switch (data->stopBits) {
   case SERIALPORT_STOPBITS_ONE:
     options.c_cflag &= ~CSTOPB;
     break;
@@ -330,411 +251,70 @@ int setup(int fd, OpenBaton *data) {
     break;
   default:
     snprintf(data->errorString, sizeof(data->errorString), "Invalid stop bits setting %d", data->stopBits);
-    close(fd);
     return -1;
   }
 
-  options.c_cflag |= CLOCAL; //ignore status lines
-  options.c_cflag |= CREAD;  //enable receiver
-  options.c_cflag |= HUPCL;  //drop DTR (i.e. hangup) on close
+  options.c_cflag |= CLOCAL;  // ignore status lines
+  options.c_cflag |= CREAD;   // enable receiver
+  if (data->hupcl) {
+    options.c_cflag |= HUPCL;  // drop DTR (i.e. hangup) on close
+  }
 
   // Raw output
   options.c_oflag = 0;
 
-  // ICANON makes partial lines not readable. It should be otional.
-  // It works with ICRNL. -Giseburt
-  options.c_lflag = 0; //ICANON;
+  // ICANON makes partial lines not readable. It should be optional.
+  // It works with ICRNL.
+  options.c_lflag = 0;  // ICANON;
+  options.c_cc[VMIN]= data->vmin;
+  options.c_cc[VTIME]= data->vtime;
 
-  options.c_cc[VMIN]= platformOptions->vmin;
-  options.c_cc[VTIME]= platformOptions->vtime;
-
-  // removed this unneeded sleep.
-  // sleep(1);
-  tcflush(fd, TCIFLUSH);
+  // Note that tcsetattr() returns success if any of the requested changes could be successfully carried out.
+  // Therefore, when making multiple changes it may be necessary to follow this call with a further call to
+  // tcgetattr() to check that all changes have been performed successfully.
+  // This also fails on OSX
   tcsetattr(fd, TCSANOW, &options);
 
-  // On OS X, starting in Tiger, we can set a custom baud rate, as follows:
-#if defined(MAC_OS_X_VERSION_10_4) && (MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_4)
-  if (baudRate == -1) {
-    speed_t speed = data->baudRate;
-    if (ioctl(fd,  IOSSIOSPEED, &speed) == -1) {
-      snprintf(data->errorString, sizeof(data->errorString), "Error %s calling ioctl( ..., IOSSIOSPEED, %ld )", strerror(errno), speed );
+  if (data->lock) {
+    if (-1 == flock(fd, LOCK_EX | LOCK_NB)) {
+      snprintf(data->errorString, sizeof(data->errorString), "Error %s Cannot lock port", strerror(errno));
+      return -1;
     }
   }
-#endif
+
+  // Copy the connection options into the ConnectionOptionsBaton to set the baud rate
+  ConnectionOptionsBaton* connectionOptions = new ConnectionOptionsBaton();
+  memset(connectionOptions, 0, sizeof(ConnectionOptionsBaton));
+  connectionOptions->fd = fd;
+  connectionOptions->baudRate = data->baudRate;
+
+  if (-1 == setBaudRate(connectionOptions)) {
+    strncpy(data->errorString, connectionOptions->errorString, sizeof(data->errorString));
+    delete(connectionOptions);
+    return -1;
+  }
+  delete(connectionOptions);
+
+  // flush all unread and wrote data up to this point because it could have been received or sent with bad settings
+  // Not needed since setBaudRate does this for us
+  // tcflush(fd, TCIOFLUSH);
 
   return 1;
-
-}
-
-void EIO_Write(uv_work_t* req) {
-  QueuedWrite* queuedWrite = static_cast<QueuedWrite*>(req->data);
-  WriteBaton* data = static_cast<WriteBaton*>(queuedWrite->baton);
-
-  data->result = 0;
-  errno = 0;
-
-  // We carefully *DON'T* break out of this loop.
-  do {
-    if ((data->result = write(data->fd, data->bufferData + data->offset, data->bufferLength - data->offset)) == -1) {
-      if (errno == EAGAIN || errno == EWOULDBLOCK)
-        return;
-
-      // The write call might be interrupted, if it is we just try again immediately.
-      if (errno != EINTR) {
-        snprintf(data->errorString, sizeof(data->errorString), "Error %s calling write(...)", strerror(errno) );
-        return;
-      }
-
-      // try again...
-      continue;
-    }
-    // there wasn't an error, do the math on what we actually wrote...
-    else {
-      data->offset += data->result;
-    }
-
-    // if we get there, we really don't want to loop
-    // break;
-  } while (data->bufferLength > data->offset);
 }
 
 void EIO_Close(uv_work_t* req) {
-  CloseBaton* data = static_cast<CloseBaton*>(req->data);
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
 
-  // printf(">>>> close fd %d\n", data->fd);
-
-  // fcntl(data->fd, F_SETFL, FNONBLOCK);
-
-  ssize_t r;
-
-  r = close(data->fd);
-
-  // printf(">>>> closed fd %d (err: %d)\n", data->fd, errno);
-
-  if (r && r != EBADF)
-    snprintf(data->errorString, sizeof(data->errorString), "Unable to close fd %d, errno: %d", data->fd, errno);
-}
-
-#ifdef __APPLE__
-
-// Function prototypes
-static kern_return_t FindModems(io_iterator_t *matchingServices);
-static io_service_t GetUsbDevice(io_service_t service);
-static stDeviceListItem* GetSerialDevices();
-
-
-static kern_return_t FindModems(io_iterator_t *matchingServices)
-{
-    kern_return_t     kernResult;
-    CFMutableDictionaryRef  classesToMatch;
-    classesToMatch = IOServiceMatching(kIOSerialBSDServiceValue);
-    if (classesToMatch != NULL)
-    {
-        CFDictionarySetValue(classesToMatch,
-                             CFSTR(kIOSerialBSDTypeKey),
-                             CFSTR(kIOSerialBSDAllTypes));
-    }
-
-    kernResult = IOServiceGetMatchingServices(kIOMasterPortDefault, classesToMatch, matchingServices);
-
-    return kernResult;
-}
-
-static io_service_t GetUsbDevice(io_service_t service)
-{
-  IOReturn status;
-  io_iterator_t   iterator = 0;
-  io_service_t    device = 0;
-
-  if (!service) {
-    return device;
+  if (-1 == close(data->fd)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, unable to close fd %d", strerror(errno), data->fd);
   }
-
-  status = IORegistryEntryCreateIterator(service,
-                                         kIOServicePlane,
-                                         (kIORegistryIterateParents | kIORegistryIterateRecursively),
-                                         &iterator);
-
-  if (status == kIOReturnSuccess)
-  {
-    io_service_t currentService;
-    while ((currentService = IOIteratorNext(iterator)) && device == 0)
-    {
-      io_name_t serviceName;
-      status = IORegistryEntryGetNameInPlane(currentService, kIOServicePlane, serviceName);
-      if (status == kIOReturnSuccess && IOObjectConformsTo(currentService, kIOUSBDeviceClassName)) {
-        device = currentService;
-      }
-      else {
-        // Release the service object which is no longer needed
-        (void) IOObjectRelease(currentService);
-      }
-    }
-
-    // Release the iterator
-    (void) IOObjectRelease(iterator);
-  }
-
-  return device;
-}
-
-static void ExtractUsbInformation(stSerialDevice *serialDevice, IOUSBDeviceInterface  **deviceInterface)
-{
-    kern_return_t kernResult;
-    UInt32 locationID;
-    kernResult = (*deviceInterface)->GetLocationID(deviceInterface, &locationID);
-    if (KERN_SUCCESS == kernResult)
-    {
-        snprintf(serialDevice->locationId, 11, "0x%08x", locationID);
-    }
-
-    UInt16 vendorID;
-    kernResult = (*deviceInterface)->GetDeviceVendor(deviceInterface, &vendorID);
-    if (KERN_SUCCESS == kernResult)
-    {
-        snprintf(serialDevice->vendorId, 7, "0x%04x", vendorID);
-    }
-
-    UInt16 productID;
-    kernResult = (*deviceInterface)->GetDeviceProduct(deviceInterface, &productID);
-    if (KERN_SUCCESS == kernResult)
-    {
-        snprintf(serialDevice->productId, 7, "0x%04x", productID);
-    }
-}
-
-static stDeviceListItem* GetSerialDevices()
-{
-    kern_return_t kernResult;
-    io_iterator_t serialPortIterator;
-    char bsdPath[MAXPATHLEN];
-
-    FindModems(&serialPortIterator);
-
-    io_service_t modemService;
-    kernResult = KERN_FAILURE;
-    Boolean modemFound = false;
-
-    // Initialize the returned path
-    *bsdPath = '\0';
-
-    stDeviceListItem* devices = NULL;
-    stDeviceListItem* lastDevice = NULL;
-    int length = 0;
-
-    while ((modemService = IOIteratorNext(serialPortIterator)))
-    {
-        CFTypeRef bsdPathAsCFString;
-
-        bsdPathAsCFString = IORegistryEntrySearchCFProperty(modemService, kIOServicePlane, CFSTR(kIOCalloutDeviceKey), kCFAllocatorDefault, kIORegistryIterateRecursively);
-
-        if (bsdPathAsCFString)
-        {
-            Boolean result;
-
-            // Convert the path from a CFString to a C (NUL-terminated)
-
-            result = CFStringGetCString((CFStringRef) bsdPathAsCFString,
-                                        bsdPath,
-                                        sizeof(bsdPath),
-                                        kCFStringEncodingUTF8);
-            CFRelease(bsdPathAsCFString);
-
-            if (result)
-            {
-                stDeviceListItem *deviceListItem = (stDeviceListItem*) malloc(sizeof(stDeviceListItem));
-                stSerialDevice *serialDevice = &(deviceListItem->value);
-                strcpy(serialDevice->port, bsdPath);
-                memset(serialDevice->locationId, 0, sizeof(serialDevice->locationId));
-                memset(serialDevice->vendorId, 0, sizeof(serialDevice->vendorId));
-                memset(serialDevice->productId, 0, sizeof(serialDevice->productId));
-                serialDevice->manufacturer[0] = '\0';
-                serialDevice->serialNumber[0] = '\0';
-                deviceListItem->next = NULL;
-                deviceListItem->length = &length;
-
-                if (devices == NULL) {
-                    devices = deviceListItem;
-                }
-                else {
-                    lastDevice->next = deviceListItem;
-                }
-
-                lastDevice = deviceListItem;
-                length++;
-
-                modemFound = true;
-                kernResult = KERN_SUCCESS;
-
-                uv_mutex_lock(&list_mutex);
-
-                io_service_t device = GetUsbDevice(modemService);
-
-                if (device) {
-                    CFStringRef manufacturerAsCFString = (CFStringRef) IORegistryEntryCreateCFProperty(device,
-                                          CFSTR(kUSBVendorString),
-                                          kCFAllocatorDefault,
-                                          0);
-
-                    if (manufacturerAsCFString)
-                    {
-                        Boolean result;
-                        char    manufacturer[MAXPATHLEN];
-
-                        // Convert from a CFString to a C (NUL-terminated)
-                        result = CFStringGetCString(manufacturerAsCFString,
-                                                    manufacturer,
-                                                    sizeof(manufacturer),
-                                                    kCFStringEncodingUTF8);
-
-                        if (result) {
-                          strcpy(serialDevice->manufacturer, manufacturer);
-                        }
-
-                        CFRelease(manufacturerAsCFString);
-                    }
-
-                    CFStringRef serialNumberAsCFString = (CFStringRef) IORegistryEntrySearchCFProperty(device,
-                                          kIOServicePlane,
-                                          CFSTR(kUSBSerialNumberString),
-                                          kCFAllocatorDefault,
-                                          kIORegistryIterateRecursively);
-
-                    if (serialNumberAsCFString)
-                    {
-                        Boolean result;
-                        char    serialNumber[MAXPATHLEN];
-
-                        // Convert from a CFString to a C (NUL-terminated)
-                        result = CFStringGetCString(serialNumberAsCFString,
-                                                    serialNumber,
-                                                    sizeof(serialNumber),
-                                                    kCFStringEncodingUTF8);
-
-                        if (result) {
-                          strcpy(serialDevice->serialNumber, serialNumber);
-                        }
-
-                        CFRelease(serialNumberAsCFString);
-                    }
-
-                    IOCFPlugInInterface **plugInInterface = NULL;
-                    SInt32        score;
-                    HRESULT       res;
-
-                    IOUSBDeviceInterface  **deviceInterface = NULL;
-
-                    kernResult = IOCreatePlugInInterfaceForService(device, kIOUSBDeviceUserClientTypeID, kIOCFPlugInInterfaceID,
-                                                           &plugInInterface, &score);
-
-                    if ((kIOReturnSuccess != kernResult) || !plugInInterface) {
-                        continue;
-                    }
-
-                    // Use the plugin interface to retrieve the device interface.
-                    res = (*plugInInterface)->QueryInterface(plugInInterface, CFUUIDGetUUIDBytes(kIOUSBDeviceInterfaceID),
-                                                             (LPVOID*) &deviceInterface);
-
-                    // Now done with the plugin interface.
-                    (*plugInInterface)->Release(plugInInterface);
-
-                    if (res || deviceInterface == NULL) {
-                        continue;
-                    }
-
-                    // Extract the desired Information
-                    ExtractUsbInformation(serialDevice, deviceInterface);
-
-                    // Release the Interface
-                    (*deviceInterface)->Release(deviceInterface);
-
-                    // Release the device
-                    (void) IOObjectRelease(device);
-                }
-
-                uv_mutex_unlock(&list_mutex);
-            }
-        }
-
-        // Release the io_service_t now that we are done with it.
-        (void) IOObjectRelease(modemService);
-    }
-
-    IOObjectRelease(serialPortIterator);  // Release the iterator.
-
-    return devices;
-}
-
-#endif
-
-void EIO_List(uv_work_t* req) {
-  // This code exists in javascript for unix platforms
-
-#ifdef __APPLE__
-  if(!lockInitialised)
-  {
-    uv_mutex_init(&list_mutex);
-    lockInitialised = TRUE;
-  }
-
-  ListBaton* data = static_cast<ListBaton*>(req->data);
-
-  stDeviceListItem* devices = GetSerialDevices();
-
-  if (*(devices->length) > 0)
-  {
-    stDeviceListItem* next = devices;
-
-    for (int i = 0, len = *(devices->length); i < len; i++) {
-        stSerialDevice device = (* next).value;
-
-        ListResultItem* resultItem = new ListResultItem();
-        resultItem->comName = device.port;
-
-        if (*device.locationId) {
-          resultItem->locationId = device.locationId;
-        }
-        if (*device.vendorId) {
-          resultItem->vendorId = device.vendorId;
-        }
-        if (*device.productId) {
-          resultItem->productId = device.productId;
-        }
-        if (*device.manufacturer) {
-          resultItem->manufacturer = device.manufacturer;
-        }
-        if (*device.serialNumber) {
-          resultItem->serialNumber = device.serialNumber;
-        }
-        data->results.push_back(resultItem);
-
-        stDeviceListItem* current = next;
-
-        if (next->next != NULL)
-        {
-          next = next->next;
-        }
-
-        free(current);
-    }
-
-  }
-
-#endif
-}
-
-void EIO_Flush(uv_work_t* req) {
-  FlushBaton* data = static_cast<FlushBaton*>(req->data);
-
-  data->result = tcflush(data->fd, TCIFLUSH);
 }
 
 void EIO_Set(uv_work_t* req) {
   SetBaton* data = static_cast<SetBaton*>(req->data);
 
   int bits;
-  ioctl( data->fd, TIOCMGET, &bits );
+  ioctl(data->fd, TIOCMGET, &bits);
 
   bits &= ~(TIOCM_RTS | TIOCM_CTS | TIOCM_DTR | TIOCM_DSR);
 
@@ -754,20 +334,52 @@ void EIO_Set(uv_work_t* req) {
     bits |= TIOCM_DSR;
   }
 
-  //todo check these returns
+  int result = 0;
   if (data->brk) {
-    ioctl(data->fd, TIOCSBRK, NULL);
-  }else{
-    ioctl(data->fd, TIOCCBRK, NULL);
+    result = ioctl(data->fd, TIOCSBRK, NULL);
+  } else {
+    result = ioctl(data->fd, TIOCCBRK, NULL);
   }
 
-  data->result = ioctl(data->fd, TIOCMSET, &bits );
+  if (-1 == result) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot set", strerror(errno));
+    return;
+  }
+
+  if (-1 == ioctl(data->fd, TIOCMSET, &bits)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot set", strerror(errno));
+    return;
+  }
+}
+
+void EIO_Get(uv_work_t* req) {
+  GetBaton* data = static_cast<GetBaton*>(req->data);
+
+  int bits;
+  if (-1 == ioctl(data->fd, TIOCMGET, &bits)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot get", strerror(errno));
+    return;
+  }
+
+  data->cts = bits & TIOCM_CTS;
+  data->dsr = bits & TIOCM_DSR;
+  data->dcd = bits & TIOCM_CD;
+}
+
+void EIO_Flush(uv_work_t* req) {
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
+
+  if (-1 == tcflush(data->fd, TCIOFLUSH)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot flush", strerror(errno));
+    return;
+  }
 }
 
 void EIO_Drain(uv_work_t* req) {
-  DrainBaton* data = static_cast<DrainBaton*>(req->data);
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
 
-  data->result = tcdrain(data->fd);
+  if (-1 == tcdrain(data->fd)) {
+    snprintf(data->errorString, sizeof(data->errorString), "Error: %s, cannot drain", strerror(errno));
+    return;
+  }
 }
-
-#endif
