@@ -1,51 +1,39 @@
-#include "serialport.h"
-#include <list>
-#include "win/disphelper.h"
-
-#include "win/stdafx.h"
-#include "win/enumser.h"
-
+#include "./serialport.h"
+#include "./serialport_win.h"
 #include <nan.h>
-
-#ifdef WIN32
+#include <list>
+#include <vector>
+#include <string.h>
+#include <windows.h>
+#include <Setupapi.h>
+#include <devguid.h>
+#pragma comment (lib, "setupapi.lib")
 
 #define MAX_BUFFER_SIZE 1000
 #define ASCII_XOFF 19
 #define ASCII_XON 17
 
-
-struct WindowsPlatformOptions : OpenBatonPlatformOptions
-{
-};
-
-OpenBatonPlatformOptions* ParsePlatformOptions(const v8::Local<v8::Object>& options){
-  // currently none
-  return new WindowsPlatformOptions();
-}
-
-
 // Declare type of pointer to CancelIoEx function
 typedef BOOL (WINAPI *CancelIoExType)(HANDLE hFile, LPOVERLAPPED lpOverlapped);
 
-
 std::list<int> g_closingHandles;
-int bufferSize;
+
 void ErrorCodeToString(const char* prefix, int errorCode, char *errorStr) {
-  switch(errorCode) {
+  switch (errorCode) {
   case ERROR_FILE_NOT_FOUND:
-    _snprintf(errorStr, ERROR_STRING_SIZE, "%s: File not found", prefix);
+    _snprintf_s(errorStr, ERROR_STRING_SIZE, _TRUNCATE, "%s: File not found", prefix);
     break;
   case ERROR_INVALID_HANDLE:
-    _snprintf(errorStr, ERROR_STRING_SIZE, "%s: Invalid handle", prefix);
+    _snprintf_s(errorStr, ERROR_STRING_SIZE, _TRUNCATE, "%s: Invalid handle", prefix);
     break;
   case ERROR_ACCESS_DENIED:
-    _snprintf(errorStr, ERROR_STRING_SIZE, "%s: Access denied", prefix);
+    _snprintf_s(errorStr, ERROR_STRING_SIZE, _TRUNCATE, "%s: Access denied", prefix);
     break;
   case ERROR_OPERATION_ABORTED:
-    _snprintf(errorStr, ERROR_STRING_SIZE, "%s: operation aborted", prefix);
+    _snprintf_s(errorStr, ERROR_STRING_SIZE, _TRUNCATE, "%s: Operation aborted", prefix);
     break;
   default:
-    _snprintf(errorStr, ERROR_STRING_SIZE, "%s: Unknown error code %d", prefix, errorCode);
+    _snprintf_s(errorStr, ERROR_STRING_SIZE, _TRUNCATE, "%s: Unknown error code %d", prefix, errorCode);
     break;
   }
 }
@@ -53,55 +41,84 @@ void ErrorCodeToString(const char* prefix, int errorCode, char *errorStr) {
 void EIO_Open(uv_work_t* req) {
   OpenBaton* data = static_cast<OpenBaton*>(req->data);
 
+  char originalPath[1024];
+  strncpy_s(originalPath, sizeof(originalPath), data->path, _TRUNCATE);
   // data->path is char[1024] but on Windows it has the form "COMx\0" or "COMxx\0"
   // We want to prepend "\\\\.\\" to it before we call CreateFile
   strncpy(data->path + 20, data->path, 10);
   strncpy(data->path, "\\\\.\\", 4);
   strncpy(data->path + 4, data->path + 20, 10);
 
+  int shareMode = FILE_SHARE_READ | FILE_SHARE_WRITE;
+  if (data->lock) {
+    shareMode = 0;
+  }
+
   HANDLE file = CreateFile(
     data->path,
     GENERIC_READ | GENERIC_WRITE,
-    0,
+    shareMode,  // dwShareMode 0 Prevents other processes from opening if they request delete, read, or write access
     NULL,
     OPEN_EXISTING,
-    FILE_FLAG_OVERLAPPED,
-    NULL);
+    FILE_FLAG_OVERLAPPED,  // allows for reading and writing at the same time and sets the handle for asynchronous I/O
+    NULL
+  );
+
   if (file == INVALID_HANDLE_VALUE) {
     DWORD errorCode = GetLastError();
     char temp[100];
-    _snprintf(temp, sizeof(temp), "Opening %s", data->path);
+    _snprintf_s(temp, sizeof(temp), _TRUNCATE, "Opening %s", originalPath);
     ErrorCodeToString(temp, errorCode, data->errorString);
     return;
   }
 
-  bufferSize = data->bufferSize;
-  if(bufferSize > MAX_BUFFER_SIZE) {
-    bufferSize = MAX_BUFFER_SIZE;
-  }
-
   DCB dcb = { 0 };
+  SecureZeroMemory(&dcb, sizeof(DCB));
   dcb.DCBlength = sizeof(DCB);
-  if(data->hupcl == false) {
-      dcb.fDtrControl = DTR_CONTROL_DISABLE; // disable DTR to avoid reset
-  } else {
-    dcb.fDtrControl = DTR_CONTROL_ENABLE;
+
+  if (!GetCommState(file, &dcb)) {
+    ErrorCodeToString("Open (GetCommState)", GetLastError(), data->errorString);
+    CloseHandle(file);
+    return;
   }
 
-  dcb.BaudRate = CBR_9600;
+  if (data->hupcl) {
+    dcb.fDtrControl = DTR_CONTROL_ENABLE;
+  } else {
+    dcb.fDtrControl = DTR_CONTROL_DISABLE;  // disable DTR to avoid reset
+  }
+
   dcb.Parity = NOPARITY;
   dcb.ByteSize = 8;
   dcb.StopBits = ONESTOPBIT;
-  dcb.fInX = FALSE;
-  dcb.fOutX = FALSE;
+
+
   dcb.fOutxDsrFlow = FALSE;
   dcb.fOutxCtsFlow = FALSE;
-  dcb.fRtsControl = RTS_CONTROL_ENABLE;
+
+  if (data->xon) {
+    dcb.fOutX = TRUE;
+  } else {
+    dcb.fOutX = FALSE;
+  }
+
+  if (data->xoff) {
+    dcb.fInX = TRUE;
+  } else {
+    dcb.fInX = FALSE;
+  }
+
+  if (data->rtscts) {
+    dcb.fRtsControl = RTS_CONTROL_ENABLE;
+  } else {
+    dcb.fRtsControl = RTS_CONTROL_DISABLE;
+  }
 
   dcb.fBinary = true;
   dcb.BaudRate = data->baudRate;
   dcb.ByteSize = data->dataBits;
-  switch(data->parity) {
+
+  switch (data->parity) {
   case SERIALPORT_PARITY_NONE:
     dcb.Parity = NOPARITY;
     break;
@@ -118,7 +135,8 @@ void EIO_Open(uv_work_t* req) {
     dcb.Parity = SPACEPARITY;
     break;
   }
-  switch(data->stopBits) {
+
+  switch (data->stopBits) {
   case SERIALPORT_STOPBITS_ONE:
     dcb.StopBits = ONESTOPBIT;
     break;
@@ -165,29 +183,24 @@ void EIO_Open(uv_work_t* req) {
 	dcb.fDtrControl = DTR_CONTROL_HANDSHAKE; 
   }
 
-  if(!SetCommState(file, &dcb)) {
-    ErrorCodeToString("SetCommState", GetLastError(), data->errorString);
+  if (!SetCommState(file, &dcb)) {
+    ErrorCodeToString("Open (SetCommState)", GetLastError(), data->errorString);
+    CloseHandle(file);
     return;
   }
 
-  // Set the com port read/write timeouts
-  DWORD serialBitsPerByte = 8/*std data bits*/ + 1/*start bit*/;
-  serialBitsPerByte += (data->parity   == SERIALPORT_PARITY_NONE ) ? 0 : 1;
-  serialBitsPerByte += (data->stopBits == SERIALPORT_STOPBITS_ONE) ? 1 : 2;
-  DWORD msPerByte = (data->baudRate > 0) ?
-                    ((1000 * serialBitsPerByte + data->baudRate - 1) / data->baudRate) :
-                    1;
-  if (msPerByte < 1) {
-    msPerByte = 1;
-  }
-  COMMTIMEOUTS commTimeouts = {0};
-  commTimeouts.ReadIntervalTimeout = msPerByte; // Minimize chance of concatenating of separate serial port packets on read
-  commTimeouts.ReadTotalTimeoutMultiplier  = 0; // Do not allow big read timeout when big read buffer used
-  commTimeouts.ReadTotalTimeoutConstant    = 1000; // Total read timeout (period of read loop)
-  commTimeouts.WriteTotalTimeoutConstant   = 1000; // Const part of write timeout
-  commTimeouts.WriteTotalTimeoutMultiplier = msPerByte; // Variable part of write timeout (per byte)
-  if(!SetCommTimeouts(file, &commTimeouts)) {
-    ErrorCodeToString("SetCommTimeouts", GetLastError(), data->errorString);
+  // Set the timeouts for read and write operations.
+  // Read operation will wait for at least 1 byte to be received.
+  COMMTIMEOUTS commTimeouts = {};
+  commTimeouts.ReadIntervalTimeout = 0;         // Never timeout, always wait for data.
+  commTimeouts.ReadTotalTimeoutMultiplier = 0;  // Do not allow big read timeout when big read buffer used
+  commTimeouts.ReadTotalTimeoutConstant = 0;    // Total read timeout (period of read loop)
+  commTimeouts.WriteTotalTimeoutConstant = 0;   // Const part of write timeout
+  commTimeouts.WriteTotalTimeoutMultiplier = 0; // Variable part of write timeout (per byte)
+
+  if (!SetCommTimeouts(file, &commTimeouts)) {
+    ErrorCodeToString("Open (SetCommTimeouts)", GetLastError(), data->errorString);
+    CloseHandle(file);
     return;
   }
 
@@ -198,42 +211,44 @@ void EIO_Open(uv_work_t* req) {
   data->result = (int)file;
 }
 
-struct WatchPortBaton {
-public:
-  HANDLE fd;
-  DWORD bytesRead;
-  char buffer[MAX_BUFFER_SIZE];
-  char errorString[ERROR_STRING_SIZE];
-  DWORD errorCode;
-  bool disconnected;
-  Nan::Callback* dataCallback;
-  Nan::Callback* errorCallback;
-  Nan::Callback* disconnectedCallback;
-};
-
 void EIO_Update(uv_work_t* req) {
+  ConnectionOptionsBaton* data = static_cast<ConnectionOptionsBaton*>(req->data);
 
+  DCB dcb = { 0 };
+  SecureZeroMemory(&dcb, sizeof(DCB));
+  dcb.DCBlength = sizeof(DCB);
+
+  if (!GetCommState((HANDLE)data->fd, &dcb)) {
+    ErrorCodeToString("Update (GetCommState)", GetLastError(), data->errorString);
+    return;
+  }
+
+  dcb.BaudRate = data->baudRate;
+
+  if (!SetCommState((HANDLE)data->fd, &dcb)) {
+    ErrorCodeToString("Update (SetCommState)", GetLastError(), data->errorString);
+    return;
+  }
 }
-
 
 void EIO_Set(uv_work_t* req) {
   SetBaton* data = static_cast<SetBaton*>(req->data);
 
   if (data->rts) {
     EscapeCommFunction((HANDLE)data->fd, SETRTS);
-  }else{
+  } else {
     EscapeCommFunction((HANDLE)data->fd, CLRRTS);
   }
 
   if (data->dtr) {
     EscapeCommFunction((HANDLE)data->fd, SETDTR);
-  }else{
+  } else {
     EscapeCommFunction((HANDLE)data->fd, CLRDTR);
   }
 
   if (data->brk) {
     EscapeCommFunction((HANDLE)data->fd, SETBREAK);
-  }else{
+  } else {
     EscapeCommFunction((HANDLE)data->fd, CLRBREAK);
   }
 
@@ -241,7 +256,7 @@ void EIO_Set(uv_work_t* req) {
 
   GetCommMask((HANDLE)data->fd, &bits);
 
-  bits &= ~( EV_CTS | EV_DSR);
+  bits &= ~(EV_CTS | EV_DSR);
 
   if (data->cts) {
     bits |= EV_CTS;
@@ -251,78 +266,29 @@ void EIO_Set(uv_work_t* req) {
     bits |= EV_DSR;
   }
 
-  data->result = SetCommMask((HANDLE)data->fd, bits);
+  if (!SetCommMask((HANDLE)data->fd, bits)) {
+    ErrorCodeToString("Setting options on COM port (SetCommMask)", GetLastError(), data->errorString);
+    return;
+  }
 }
 
+void EIO_Get(uv_work_t* req) {
+  GetBaton* data = static_cast<GetBaton*>(req->data);
 
-void EIO_WatchPort(uv_work_t* req) {
-  WatchPortBaton* data = static_cast<WatchPortBaton*>(req->data);
-  data->bytesRead = 0;
-  data->disconnected = false;
-
-  // Event used by GetOverlappedResult(..., TRUE) to wait for incoming data or timeout
-  // Event MUST be used if program has several simultaneous asynchronous operations
-  // on the same handle (i.e. ReadFile and WriteFile)
-  HANDLE hEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
-
-  while(true) {
-    OVERLAPPED ov = {0};
-    ov.hEvent = hEvent;
-
-    // Start read operation - synchrounous or asynchronous
-    DWORD bytesReadSync = 0;
-    if(!ReadFile((HANDLE)data->fd, data->buffer, bufferSize, &bytesReadSync, &ov)) {
-      data->errorCode = GetLastError();
-      if(data->errorCode != ERROR_IO_PENDING) {
-        // Read operation error
-        if(data->errorCode == ERROR_OPERATION_ABORTED) {
-          data->disconnected = true;
-        }
-        else {
-          ErrorCodeToString("Reading from COM port (ReadFile)", data->errorCode, data->errorString);
-        }
-        break;
-      }
-
-      // Read operation is asynchronous and is pending
-      // We MUST wait for operation completion before deallocation of OVERLAPPED struct
-      // or read data buffer
-
-      // Wait for async read operation completion or timeout
-      DWORD bytesReadAsync = 0;
-      if(!GetOverlappedResult((HANDLE)data->fd, &ov, &bytesReadAsync, TRUE)) {
-        // Read operation error
-        data->errorCode = GetLastError();
-        if(data->errorCode == ERROR_OPERATION_ABORTED) {
-          data->disconnected = true;
-        }
-        else {
-          ErrorCodeToString("Reading from COM port (GetOverlappedResult)", data->errorCode, data->errorString);
-        }
-        break;
-      }
-      else {
-        // Read operation completed asynchronously
-        data->bytesRead = bytesReadAsync;
-      }
-    }
-    else {
-      // Read operation completed synchronously
-      data->bytesRead = bytesReadSync;
-    }
-
-    // Return data received if any
-    if(data->bytesRead > 0) {
-      break;
-    }
+  DWORD bits = 0;
+  if (!GetCommModemStatus((HANDLE)data->fd, &bits)) {
+    ErrorCodeToString("Getting control settings on COM port (GetCommModemStatus)", GetLastError(), data->errorString);
+    return;
   }
 
-  CloseHandle(hEvent);
+  data->cts = bits & MS_CTS_ON;
+  data->dsr = bits & MS_DSR_ON;
+  data->dcd = bits & MS_RLSD_ON;
 }
 
 bool IsClosingHandle(int fd) {
-  for(std::list<int>::iterator it=g_closingHandles.begin(); it!=g_closingHandles.end(); ++it) {
-    if(fd == *it) {
+  for (std::list<int>::iterator it = g_closingHandles.begin(); it != g_closingHandles.end(); ++it) {
+    if (fd == *it) {
       g_closingHandles.remove(fd);
       return true;
     }
@@ -330,123 +296,287 @@ bool IsClosingHandle(int fd) {
   return false;
 }
 
-void DisposeWatchPortCallbacks(WatchPortBaton* data) {
-  delete data->dataCallback;
-  delete data->errorCallback;
-  delete data->disconnectedCallback;
-}
+NAN_METHOD(Write) {
+  // file descriptor
+  if (!info[0]->IsInt32()) {
+    Nan::ThrowTypeError("First argument must be an int");
+    return;
+  }
+  int fd = Nan::To<int>(info[0]).FromJust();
 
-// FinalizerCallback will prevent WatchPortBaton::buffer from getting
-// collected by gc while finalizing v8::ArrayBuffer. The buffer will
-// get cleaned up through this callback.
-static void FinalizerCallback(char* data, void* hint) {
-  uv_work_t* req = reinterpret_cast<uv_work_t*>(hint);
-  WatchPortBaton* wpb = static_cast<WatchPortBaton*>(req->data);
-  delete wpb;
-  delete req;
-}
+  // buffer
+  if (!info[1]->IsObject() || !node::Buffer::HasInstance(info[1])) {
+    Nan::ThrowTypeError("Second argument must be a buffer");
+    return;
+  }
+  v8::Local<v8::Object> buffer = info[1]->ToObject();
+  char* bufferData = node::Buffer::Data(buffer);
+  size_t bufferLength = node::Buffer::Length(buffer);
 
-void EIO_AfterWatchPort(uv_work_t* req) {
-  Nan::HandleScope scope;
-
-  WatchPortBaton* data = static_cast<WatchPortBaton*>(req->data);
-  if(data->disconnected) {
-    data->disconnectedCallback->Call(0, NULL);
-    DisposeWatchPortCallbacks(data);
-    goto cleanup;
+  // callback
+  if (!info[2]->IsFunction()) {
+    Nan::ThrowTypeError("Third argument must be a function");
+    return;
   }
 
-  bool skipCleanup = false;
-  if(data->bytesRead > 0) {
-    v8::Local<v8::Value> argv[1];
-    argv[0] = Nan::NewBuffer(data->buffer, data->bytesRead, FinalizerCallback, req).ToLocalChecked();
-    skipCleanup = true;
-    data->dataCallback->Call(1, argv);
-  } else if(data->errorCode > 0) {
-    if(data->errorCode == ERROR_INVALID_HANDLE && IsClosingHandle((int)data->fd)) {
-      DisposeWatchPortCallbacks(data);
-      goto cleanup;
-    } else {
-      v8::Local<v8::Value> argv[1];
-      argv[0] = Nan::Error(data->errorString);
-      data->errorCallback->Call(1, argv);
-      Sleep(100); // prevent the errors from occurring too fast
+  WriteBaton* baton = new WriteBaton();
+  memset(baton, 0, sizeof(WriteBaton));
+  baton->fd = fd;
+  baton->buffer.Reset(buffer);
+  baton->bufferData = bufferData;
+  baton->bufferLength = bufferLength;
+  baton->offset = 0;
+  baton->callback.Reset(info[2].As<v8::Function>());
+  baton->complete = false;
+  // WriteFileEx requires a thread that can block. Create a new thread to
+  // run the write operation, saving the handle so it can be deallocated later.
+  baton->hThread = CreateThread(NULL, 0, WriteThread, baton, 0, NULL);
+}
+
+void __stdcall WriteIOCompletion(DWORD errorCode, DWORD bytesTransferred, OVERLAPPED* ov) {
+  WriteBaton* baton = static_cast<WriteBaton*>(ov->hEvent);
+  DWORD bytesWritten;
+  if (!GetOverlappedResult((HANDLE)baton->fd, ov, &bytesWritten, TRUE)) {
+    errorCode = GetLastError();
+    ErrorCodeToString("Writing to COM port (GetOverlappedResult)", errorCode, baton->errorString);
+    baton->complete = true;
+    return;
+  }
+  if (bytesWritten) {
+    baton->offset += bytesWritten;
+    if (baton->offset >= baton->bufferLength) {
+      baton->complete = true;
     }
   }
-  AfterOpenSuccess((int)data->fd, data->dataCallback, data->disconnectedCallback, data->errorCallback);
+}
 
-cleanup:
-  if (!skipCleanup) {
-    delete data;
-    delete req;
+DWORD __stdcall WriteThread(LPVOID param) {
+  WriteBaton* baton = static_cast<WriteBaton*>(param);
+
+  OVERLAPPED* ov = new OVERLAPPED;
+  memset(ov, 0, sizeof(ov));
+  ov->hEvent = static_cast<void*>(baton);
+
+  while (!baton->complete) {
+    char* offsetPtr = baton->bufferData + baton->offset;
+    // WriteFileEx requires calling GetLastError even upon success. Clear the error beforehand.
+    SetLastError(0);
+    WriteFileEx((HANDLE)baton->fd, offsetPtr, static_cast<DWORD>(baton->bufferLength), ov, WriteIOCompletion);
+    // Error codes when call is successful, such as ERROR_MORE_DATA.
+    DWORD lastError = GetLastError();
+    if (lastError != ERROR_SUCCESS) {
+      ErrorCodeToString("Writing to COM port (WriteFileEx)", lastError, baton->errorString);
+      break;
+    }
+    // IOCompletion routine is only called once this thread is in an alertable wait state.
+    SleepEx(INFINITE, TRUE);
   }
+  delete ov;
+  // Signal the main thread to run the callback.
+  uv_async_t* async = new uv_async_t;
+  uv_async_init(uv_default_loop(), async, EIO_AfterWrite);
+  async->data = baton;
+  uv_async_send(async);
+  ExitThread(0);
 }
 
-void AfterOpenSuccess(int fd, Nan::Callback* dataCallback, Nan::Callback* disconnectedCallback, Nan::Callback* errorCallback) {
-  WatchPortBaton* baton = new WatchPortBaton();
-  memset(baton, 0, sizeof(WatchPortBaton));
-  baton->fd = (HANDLE)fd;
-  baton->dataCallback = dataCallback;
-  baton->errorCallback = errorCallback;
-  baton->disconnectedCallback = disconnectedCallback;
+void EIO_AfterWrite(uv_async_t* req) {
+  Nan::HandleScope scope;
+  WriteBaton* baton = static_cast<WriteBaton*>(req->data);
+  WaitForSingleObject(baton->hThread, INFINITE);
+  delete req;
 
-  uv_work_t* req = new uv_work_t();
-  req->data = baton;
-
-  uv_queue_work(uv_default_loop(), req, EIO_WatchPort, (uv_after_work_cb)EIO_AfterWatchPort);
+  v8::Local<v8::Value> argv[1];
+  if (baton->errorString[0]) {
+    argv[0] = v8::Exception::Error(Nan::New<v8::String>(baton->errorString).ToLocalChecked());
+  } else {
+    argv[0] = Nan::Null();
+  }
+  baton->callback.Call(1, argv);
+  delete baton;
 }
 
-void EIO_Write(uv_work_t* req) {
-  QueuedWrite* queuedWrite = static_cast<QueuedWrite*>(req->data);
-  WriteBaton* data = static_cast<WriteBaton*>(queuedWrite->baton);
-  data->result = 0;
+NAN_METHOD(Read) {
+  // file descriptor
+  if (!info[0]->IsInt32()) {
+    Nan::ThrowTypeError("First argument must be a fd");
+    return;
+  }
+  int fd = Nan::To<int>(info[0]).FromJust();
 
-  do {
-	  OVERLAPPED ov = {0};
-	  // Event used by GetOverlappedResult(..., TRUE) to wait for outgoing data or timeout
-	  // Event MUST be used if program has several simultaneous asynchronous operations
-	  // on the same handle (i.e. ReadFile and WriteFile)
-	  ov.hEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
+  // buffer
+  if (!info[1]->IsObject() || !node::Buffer::HasInstance(info[1])) {
+    Nan::ThrowTypeError("Second argument must be a buffer");
+    return;
+  }
+  v8::Local<v8::Object> buffer = info[1]->ToObject();
+  size_t bufferLength = node::Buffer::Length(buffer);
 
-	  // Start write operation - synchrounous or asynchronous
-	  DWORD bytesWrittenSync = 0;
-	  if(!WriteFile((HANDLE)data->fd, data->bufferData, static_cast<DWORD>(data->bufferLength), &bytesWrittenSync, &ov)) {
-		DWORD lastError = GetLastError();
-		if(lastError != ERROR_IO_PENDING) {
-		  // Write operation error
-		  ErrorCodeToString("Writing to COM port (WriteFile)", lastError, data->errorString);
-		}
-		else {
-		  // Write operation is asynchronous and is pending
-		  // We MUST wait for operation completion before deallocation of OVERLAPPED struct
-		  // or write data buffer
+  // offset
+  if (!info[2]->IsInt32()) {
+    Nan::ThrowTypeError("Third argument must be an int");
+    return;
+  }
+  int offset = Nan::To<v8::Int32>(info[2]).ToLocalChecked()->Value();
 
-		  // Wait for async write operation completion or timeout
-		  DWORD bytesWrittenAsync = 0;
-		  if(!GetOverlappedResult((HANDLE)data->fd, &ov, &bytesWrittenAsync, TRUE)) {
-			// Write operation error
-			DWORD lastError = GetLastError();
-			ErrorCodeToString("Writing to COM port (GetOverlappedResult)", lastError, data->errorString);
-		  }
-		  else {
-			// Write operation completed asynchronously
-			data->result = bytesWrittenAsync;
-		  }
-		}
-	  }
-	  else {
-		// Write operation completed synchronously
-		data->result = bytesWrittenSync;
-	  }
+  // bytes to read
+  if (!info[3]->IsInt32()) {
+    Nan::ThrowTypeError("Fourth argument must be an int");
+    return;
+  }
+  size_t bytesToRead = Nan::To<v8::Int32>(info[3]).ToLocalChecked()->Value();
 
-	  data->offset += data->result;
-	  CloseHandle(ov.hEvent);
-  } while (data->bufferLength > data->offset);
+  if ((bytesToRead + offset) > bufferLength) {
+    Nan::ThrowTypeError("'bytesToRead' + 'offset' cannot be larger than the buffer's length");
+    return;
+  }
 
+  // callback
+  if (!info[4]->IsFunction()) {
+    Nan::ThrowTypeError("Fifth argument must be a function");
+    return;
+  }
+
+  ReadBaton* baton = new ReadBaton();
+  memset(baton, 0, sizeof(ReadBaton));
+  baton->fd = fd;
+  baton->offset = offset;
+  baton->bytesToRead = bytesToRead;
+  baton->bufferLength = bufferLength;
+  baton->bufferData = node::Buffer::Data(buffer);
+  baton->callback.Reset(info[4].As<v8::Function>());
+  baton->complete = false;
+  // ReadFileEx requires a thread that can block. Create a new thread to
+  // run the read operation, saving the handle so it can be deallocated later.
+  baton->hThread = CreateThread(NULL, 0, ReadThread, baton, 0, NULL);
+}
+
+void __stdcall ReadIOCompletion(DWORD errorCode, DWORD bytesTransferred, OVERLAPPED* ov) {
+  ReadBaton* baton = static_cast<ReadBaton*>(ov->hEvent);
+
+  if (errorCode) {
+    ErrorCodeToString("Reading from COM port (ReadIOCompletion)", errorCode, baton->errorString);
+    baton->complete = true;
+    return;
+  }
+
+  DWORD lastError;
+  if (!GetOverlappedResult((HANDLE)baton->fd, ov, &bytesTransferred, TRUE)) {
+    lastError = GetLastError();
+    ErrorCodeToString("Reading from COM port (GetOverlappedResult)", lastError, baton->errorString);
+    baton->complete = true;
+    return;
+  }
+  if (bytesTransferred) {
+    baton->bytesRead += bytesTransferred;
+    baton->offset += bytesTransferred;
+  }
+
+  // ReadFileEx and GetOverlappedResult retrieved only 1 byte. Read any additional data in the input
+  // buffer. Set the timeout to MAXDWORD in order to disable timeouts, so the read operation will
+  // return immediately no matter how much data is available.
+  COMMTIMEOUTS commTimeouts = {};
+  commTimeouts.ReadIntervalTimeout = MAXDWORD;
+  if (!SetCommTimeouts((HANDLE)baton->fd, &commTimeouts)) {
+    lastError = GetLastError();
+    ErrorCodeToString("Setting COM timeout (SetCommTimeouts)", lastError, baton->errorString);
+    baton->complete = true;
+    return;
+  }
+
+  // Store additional data after whatever data has already been read.
+  char* offsetPtr = baton->bufferData + baton->offset;
+
+  // ReadFile, unlike ReadFileEx, needs an event in the overlapped structure.
+  ov->hEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
+  if (!ReadFile((HANDLE)baton->fd, offsetPtr, baton->bytesToRead, &bytesTransferred, ov)) {
+    errorCode = GetLastError();
+
+    if (errorCode != ERROR_IO_PENDING) {
+      ErrorCodeToString("Reading from COM port (ReadFile)", errorCode, baton->errorString);
+      baton->complete = true;
+      CloseHandle(ov->hEvent);
+      return;
+    }
+
+    if (!GetOverlappedResult((HANDLE)baton->fd, ov, &bytesTransferred, TRUE)) {
+      lastError = GetLastError();
+      ErrorCodeToString("Reading from COM port (GetOverlappedResult)", lastError, baton->errorString);
+      baton->complete = true;
+      CloseHandle(ov->hEvent);
+      return;
+    }
+  }
+  CloseHandle(ov->hEvent);
+
+  baton->bytesToRead -= bytesTransferred;
+  baton->bytesRead += bytesTransferred;
+  baton->complete = true;
+}
+
+DWORD __stdcall ReadThread(LPVOID param) {
+  ReadBaton* baton = static_cast<ReadBaton*>(param);
+  DWORD lastError;
+
+  OVERLAPPED* ov = new OVERLAPPED;
+  memset(ov, 0, sizeof(OVERLAPPED));
+  ov->hEvent = static_cast<void*>(baton);
+
+  while (!baton->complete) {
+    // Reset the read timeout to 0, so that it will block until more data arrives.
+    COMMTIMEOUTS commTimeouts = {};
+    commTimeouts.ReadIntervalTimeout = 0;
+    if (!SetCommTimeouts((HANDLE)baton->fd, &commTimeouts)) {
+      lastError = GetLastError();
+      ErrorCodeToString("Setting COM timeout (SetCommTimeouts)", lastError, baton->errorString);
+      break;
+    }
+    // ReadFileEx doesn't use overlapped's hEvent, so it is reserved for user data.
+    ov->hEvent = static_cast<HANDLE>(baton);
+    char* offsetPtr = baton->bufferData + baton->offset;
+    // ReadFileEx requires calling GetLastError even upon success. Clear the error beforehand.
+    SetLastError(0);
+    // Only read 1 byte, so that the callback will be triggered once any data arrives.
+    ReadFileEx((HANDLE)baton->fd, offsetPtr, 1, ov, ReadIOCompletion);
+    // Error codes when call is successful, such as ERROR_MORE_DATA.
+    lastError = GetLastError();
+    if (lastError != ERROR_SUCCESS) {
+      ErrorCodeToString("Reading from COM port (ReadFileEx)", lastError, baton->errorString);
+      break;
+    }
+    // IOCompletion routine is only called once this thread is in an alertable wait state.
+    SleepEx(INFINITE, TRUE);
+  }
+  delete ov;
+  // Signal the main thread to run the callback.
+  uv_async_t* async = new uv_async_t;
+  uv_async_init(uv_default_loop(), async, EIO_AfterRead);
+  async->data = baton;
+  uv_async_send(async);
+  ExitThread(0);
+}
+
+void EIO_AfterRead(uv_async_t* req) {
+  Nan::HandleScope scope;
+  ReadBaton* baton = static_cast<ReadBaton*>(req->data);
+  WaitForSingleObject(baton->hThread, INFINITE);
+  delete req;
+
+  v8::Local<v8::Value> argv[2];
+  if (baton->errorString[0]) {
+    argv[0] = Nan::Error(baton->errorString);
+    argv[1] = Nan::Undefined();
+  } else {
+    argv[0] = Nan::Null();
+    argv[1] = Nan::New<v8::Integer>((int)baton->bytesRead);
+  }
+
+  baton->callback.Call(2, argv);
+  delete baton;
 }
 
 void EIO_Close(uv_work_t* req) {
-  CloseBaton* data = static_cast<CloseBaton*>(req->data);
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
 
   g_closingHandles.push_back(data->fd);
 
@@ -454,124 +584,200 @@ void EIO_Close(uv_work_t* req) {
   // Look up function address
   CancelIoExType pCancelIoEx = (CancelIoExType)GetProcAddress(hKernel32, "CancelIoEx");
   // Do something with it
-  if (pCancelIoEx)
-  {
-      // Function exists so call it
-      // Cancel all pending IO Requests for the current device
-      pCancelIoEx((HANDLE)data->fd, NULL);
+  if (pCancelIoEx) {
+    // Function exists so call it
+    // Cancel all pending IO Requests for the current device
+    pCancelIoEx((HANDLE)data->fd, NULL);
   }
-  if(!CloseHandle((HANDLE)data->fd)) {
-    ErrorCodeToString("closing connection", GetLastError(), data->errorString);
+  if (!CloseHandle((HANDLE)data->fd)) {
+    ErrorCodeToString("Closing connection (CloseHandle)", GetLastError(), data->errorString);
     return;
   }
 }
 
-/*
- * listComPorts.c -- list COM ports
- *
- * http://github.com/todbot/usbSearch/
- *
- * 2012, Tod E. Kurt, http://todbot.com/blog/
- *
- *
- * Uses DispHealper : http://disphelper.sourceforge.net/
- *
- * Notable VIDs & PIDs combos:
- * VID 0403 - FTDI
- *
- * VID 0403 / PID 6001 - Arduino Diecimila
- *
- */
+char *copySubstring(char *someString, int n)
+{
+  char *new_ = (char*)malloc(sizeof(char)*n + 1);
+  strncpy_s(new_, n + 1, someString, n);
+  new_[n] = '\0';
+  return new_;
+}
+
+NAN_METHOD(List) {
+  // callback
+  if (!info[0]->IsFunction()) {
+    Nan::ThrowTypeError("First argument must be a function");
+    return;
+  }
+
+  ListBaton* baton = new ListBaton();
+  strcpy(baton->errorString, "");
+  baton->callback.Reset(info[0].As<v8::Function>());
+
+  uv_work_t* req = new uv_work_t();
+  req->data = baton;
+  uv_queue_work(uv_default_loop(), req, EIO_List, (uv_after_work_cb)EIO_AfterList);
+}
+
 void EIO_List(uv_work_t* req) {
   ListBaton* data = static_cast<ListBaton*>(req->data);
 
-  {
-    DISPATCH_OBJ(wmiSvc);
-    DISPATCH_OBJ(colDevices);
+  GUID *guidDev = (GUID*)& GUID_DEVCLASS_PORTS;
+  HDEVINFO hDevInfo = SetupDiGetClassDevs(guidDev, NULL, NULL, DIGCF_PRESENT | DIGCF_PROFILE);
+  SP_DEVINFO_DATA deviceInfoData;
 
-    dhInitialize(TRUE);
-    dhToggleExceptions(FALSE);
+  int memberIndex = 0;
+  DWORD dwSize, dwPropertyRegDataType;
+  char szBuffer[400];
+  char *pnpId;
+  char *vendorId;
+  char *productId;
+  char *name;
+  char *manufacturer;
+  char *locationId;
+  bool isCom;
+  while (true) {
+    pnpId = NULL;
+    vendorId = NULL;
+    productId = NULL;
+    name = NULL;
+    manufacturer = NULL;
+    locationId = NULL;
 
-    dhGetObject(L"winmgmts:{impersonationLevel=impersonate}!\\\\.\\root\\cimv2", NULL, &wmiSvc);
-    dhGetValue(L"%o", &colDevices, wmiSvc, L".ExecQuery(%S)", L"Select * from Win32_PnPEntity");
+    ZeroMemory(&deviceInfoData, sizeof(SP_DEVINFO_DATA));
+    deviceInfoData.cbSize = sizeof(SP_DEVINFO_DATA);
 
-    int port_count = 0;
-    FOR_EACH(objDevice, colDevices, NULL) {
-      char* name = NULL;
-      char* pnpid = NULL;
-      char* manu = NULL;
-      char* match;
-
-      dhGetValue(L"%s", &name,  objDevice, L".Name");
-      dhGetValue(L"%s", &pnpid, objDevice, L".PnPDeviceID");
-
-      if( name != NULL && ((match = strstr( name, "(COM" )) != NULL) ) { // look for "(COM23)"
-        // 'Manufacturuer' can be null, so only get it if we need it
-        dhGetValue(L"%s", &manu, objDevice,  L".Manufacturer");
-        port_count++;
-        char* comname = strtok( match, "()");
-        ListResultItem* resultItem = new ListResultItem();
-        resultItem->comName = comname;
-        resultItem->manufacturer = manu;
-        resultItem->pnpId = pnpid;
-        data->results.push_back(resultItem);
-        dhFreeString(manu);
-      }
-
-      dhFreeString(name);
-      dhFreeString(pnpid);
-    } NEXT(objDevice);
-
-    SAFE_RELEASE(colDevices);
-    SAFE_RELEASE(wmiSvc);
-
-    dhUninitialize(TRUE);
-  }
-
-  std::vector<UINT> ports;
-  if (CEnumerateSerial::UsingQueryDosDevice(ports))
-  {
-    for (size_t i = 0; i < ports.size(); i++)
-    {
-      char comname[64] = { 0 };
-      _snprintf(comname, sizeof(comname), "COM%u", ports[i]);
-      bool bFound = false;
-      for (std::list<ListResultItem*>::iterator ri = data->results.begin(); ri != data->results.end(); ++ri)
-      {
-        if (stricmp((*ri)->comName.c_str(), comname) == 0)
-        {
-          bFound = true;
-          break;
-        }
-      }
-      if (!bFound)
-      {
-        ListResultItem* resultItem = new ListResultItem();
-        resultItem->comName = comname;
-        resultItem->manufacturer = "";
-        resultItem->pnpId = "";
-        data->results.push_back(resultItem);
+    if (SetupDiEnumDeviceInfo(hDevInfo, memberIndex, &deviceInfoData) == FALSE) {
+      if (GetLastError() == ERROR_NO_MORE_ITEMS) {
+        break;
       }
     }
+
+    dwSize = sizeof(szBuffer);
+    SetupDiGetDeviceInstanceId(hDevInfo, &deviceInfoData, szBuffer, dwSize, &dwSize);
+    szBuffer[dwSize] = '\0';
+    pnpId = strdup(szBuffer);
+
+    vendorId = strstr(szBuffer, "VID_");
+    if (vendorId) {
+      vendorId += 4;
+      vendorId = copySubstring(vendorId, 4);
+    }
+    productId = strstr(szBuffer, "PID_");
+    if (productId) {
+      productId += 4;
+      productId = copySubstring(productId, 4);
+    }
+
+    if (SetupDiGetDeviceRegistryProperty(hDevInfo, &deviceInfoData, SPDRP_LOCATION_INFORMATION, &dwPropertyRegDataType, (BYTE*)szBuffer, sizeof(szBuffer), &dwSize)) {
+      locationId = strdup(szBuffer);
+    }
+    if (SetupDiGetDeviceRegistryProperty(hDevInfo, &deviceInfoData, SPDRP_MFG, &dwPropertyRegDataType, (BYTE*)szBuffer, sizeof(szBuffer), &dwSize)) {
+      manufacturer = strdup(szBuffer);
+    }
+
+    HKEY hkey = SetupDiOpenDevRegKey(hDevInfo, &deviceInfoData, DICS_FLAG_GLOBAL, 0, DIREG_DEV, KEY_READ);
+    if (hkey != INVALID_HANDLE_VALUE) {
+      dwSize = sizeof(szBuffer);
+      if (RegQueryValueEx(hkey, "PortName", NULL, NULL, (LPBYTE)&szBuffer, &dwSize) == ERROR_SUCCESS) {
+        szBuffer[dwSize] = '\0';
+        name = strdup(szBuffer);
+        isCom = strstr(szBuffer, "COM") != NULL;
+      }
+    }
+    if (isCom) {
+      ListResultItem* resultItem = new ListResultItem();
+      resultItem->comName = name;
+      resultItem->manufacturer = manufacturer;
+      resultItem->pnpId = pnpId;
+      if (vendorId) {
+        resultItem->vendorId = vendorId;
+      }
+      if (productId) {
+        resultItem->productId = productId;
+      }
+      if (locationId) {
+        resultItem->locationId = locationId;
+      }
+      data->results.push_back(resultItem);
+    }
+    free(pnpId);
+    free(vendorId);
+    free(productId);
+    free(locationId);
+    free(manufacturer);
+    free(name);
+
+    RegCloseKey(hkey);
+    memberIndex++;
+  }
+  if (hDevInfo) {
+    SetupDiDestroyDeviceInfoList(hDevInfo);
   }
 }
 
-void EIO_Flush(uv_work_t* req) {
-  FlushBaton* data = static_cast<FlushBaton*>(req->data);
+void setIfNotEmpty(v8::Local<v8::Object> item, std::string key, const char *value) {
+  v8::Local<v8::String> v8key = Nan::New<v8::String>(key).ToLocalChecked();
+  if (strlen(value) > 0) {
+    Nan::Set(item, v8key, Nan::New<v8::String>(value).ToLocalChecked());
+  } else {
+    Nan::Set(item, v8key, Nan::Undefined());
+  }
+}
 
-  if(!FlushFileBuffers((HANDLE)data->fd)) {
-    ErrorCodeToString("flushing connection", GetLastError(), data->errorString);
+void EIO_AfterList(uv_work_t* req) {
+  Nan::HandleScope scope;
+
+  ListBaton* data = static_cast<ListBaton*>(req->data);
+
+  v8::Local<v8::Value> argv[2];
+  if (data->errorString[0]) {
+    argv[0] = v8::Exception::Error(Nan::New<v8::String>(data->errorString).ToLocalChecked());
+    argv[1] = Nan::Undefined();
+  } else {
+    v8::Local<v8::Array> results = Nan::New<v8::Array>();
+    int i = 0;
+    for (std::list<ListResultItem*>::iterator it = data->results.begin(); it != data->results.end(); ++it, i++) {
+      v8::Local<v8::Object> item = Nan::New<v8::Object>();
+
+      setIfNotEmpty(item, "comName", (*it)->comName.c_str());
+      setIfNotEmpty(item, "manufacturer", (*it)->manufacturer.c_str());
+      setIfNotEmpty(item, "serialNumber", (*it)->serialNumber.c_str());
+      setIfNotEmpty(item, "pnpId", (*it)->pnpId.c_str());
+      setIfNotEmpty(item, "locationId", (*it)->locationId.c_str());
+      setIfNotEmpty(item, "vendorId", (*it)->vendorId.c_str());
+      setIfNotEmpty(item, "productId", (*it)->productId.c_str());
+
+      Nan::Set(results, i, item);
+    }
+    argv[0] = Nan::Null();
+    argv[1] = results;
+  }
+  data->callback.Call(2, argv);
+
+  for (std::list<ListResultItem*>::iterator it = data->results.begin(); it != data->results.end(); ++it) {
+    delete *it;
+  }
+  delete data;
+  delete req;
+}
+
+
+void EIO_Flush(uv_work_t* req) {
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
+
+  DWORD purge_all = PURGE_RXABORT | PURGE_RXCLEAR | PURGE_TXABORT | PURGE_TXCLEAR;
+  if (!PurgeComm((HANDLE)data->fd, purge_all)) {
+    ErrorCodeToString("Flushing connection (PurgeComm)", GetLastError(), data->errorString);
     return;
   }
 }
 
 void EIO_Drain(uv_work_t* req) {
-  DrainBaton* data = static_cast<DrainBaton*>(req->data);
+  VoidBaton* data = static_cast<VoidBaton*>(req->data);
 
-  if(!FlushFileBuffers((HANDLE)data->fd)) {
-    ErrorCodeToString("draining connection", GetLastError(), data->errorString);
+  if (!FlushFileBuffers((HANDLE)data->fd)) {
+    ErrorCodeToString("Draining connection (FlushFileBuffers)", GetLastError(), data->errorString);
     return;
   }
 }
-
-#endif
